Add shutdown hook; expose WS handler; cast pos
Add a JVM shutdown hook and switch Main to a non-blocking wait/logging flow (remove ENTER prompt, use 1s loop and rely on interruption/Ctrl+C) so the engine can be stopped cleanly. Change WebSocketApiServer.processConnections() visibility from protected to public to allow external invocation/testing. Cast currentPosition absolute value to int in AIEnhancedStrategy to ensure the order quantity is an int and avoid type issues.

+++ src/main/java/com/hft/strategy/AIEnhancedStrategy.java
@@ -246,7 +246,7 @@ public class AIEnhancedStrategy implements Strategy {
                 orderIdGenerator.getAndIncrement(),
                 symbolId,
                 tick.price,
-                Math.abs(currentPosition),
+                (int)Math.abs(currentPosition),
                 currentPosition > 0 ? (byte)1 : (byte)0, // Close position
                 (byte)1  // Market order
             );
